Made blockSize int and ensure that it is not initalized on scan to prevent bugs when blockSize from AppUsage used for DiskUsage

+++ src/com/google/android/diskusage/FileSystemFreeSpace.java
@@ -1,7 +1,7 @@
 package com.google.android.diskusage;
 
 public class FileSystemFreeSpace extends FileSystemSpecial {
-  public FileSystemFreeSpace(String name, long size, long blockSize) {
+  public FileSystemFreeSpace(String name, long size, int blockSize) {
     super(name, size, blockSize);
   }
 }
